+++ examples/liquidity/estimate-add-liquidity.ts
@@ -1,4 +1,5 @@
 import { Network } from '@xchainjs/xchain-client'
+import { isAssetRuneNative } from '@xchainjs/xchain-thorchain'
 import {
   AddliquidityPosition,
   CryptoAmount,
@@ -8,7 +9,7 @@ import {
   ThorchainQuery,
   Thornode,
 } from '@xchainjs/xchain-thorchain-query'
-import { assetAmount, assetFromStringEx, assetToBase, isAssetRuneNative } from '@xchainjs/xchain-util'
+import { assetAmount, assetFromStringEx, assetToBase } from '@xchainjs/xchain-util'
 
 function print(estimate: EstimateAddLP, rune: CryptoAmount, asset: CryptoAmount) {
   const expanded = {