+++ packages/xchain-bitcoincash/__tests__/utils.test.ts
@@ -1,9 +1,9 @@
 import { Network } from '@xchainjs/xchain-client'
-import { AssetBCH } from '@xchainjs/xchain-util'
 import * as bchaddr from 'bchaddrjs'
 
 import mockHaskoinApi from '../__mocks__/haskoin'
-import * as utils from '../src/utils'
+import { AssetBCH } from '../src/const'
+import * as Utils from '../src/utils'
 
 describe('Bitcoin Cash Utils Test', () => {
   beforeEach(() => {
@@ -20,46 +20,46 @@ describe('Bitcoin Cash Utils Test', () => {
   describe('stripPrefix', () => {
     it('get the right vault fee', () => {
       const memo = 'SWAP:THOR.RUNE'
-      const fee = utils.calcFee(1, memo)
+      const fee = Utils.calcFee(1, memo)
       expect(fee.amount().toNumber()).toEqual(103)
     })
 
     it('get the right normal fee', () => {
-      const fee = utils.calcFee(1)
+      const fee = Utils.calcFee(1)
       expect(fee.amount().toNumber()).toEqual(78)
     })
   })
 
   describe('stripPrefix', () => {
     it('should strip out the prefix from the testnet address', () => {
-      const strip_address = utils.stripPrefix(testnet_address)
+      const strip_address = Utils.stripPrefix(testnet_address)
       expect(strip_address).toEqual('qpd7jmj0hltgxux06v9d9u6933vq7zd0kyjlapya0g')
     })
 
     it('should strip out the prefix from the testnet address', () => {
-      const strip_address = utils.stripPrefix(mainnet_address)
+      const strip_address = Utils.stripPrefix(mainnet_address)
       expect(strip_address).toEqual('qp4kjpk684c3d9qjk5a37vl2xn86wxl0f5j2ru0daj')
     })
 
     it('should leave unprefixed address intact', () => {
-      const strip_address = utils.stripPrefix('qpd7jmj0hltgxux06v9d9u6933vq7zd0kyjlapya0g')
+      const strip_address = Utils.stripPrefix('qpd7jmj0hltgxux06v9d9u6933vq7zd0kyjlapya0g')
       expect(strip_address).toEqual('qpd7jmj0hltgxux06v9d9u6933vq7zd0kyjlapya0g')
     })
   })
 
   describe('toBCHAddressNetwork', () => {
     it('returns `bchaddr.Network.Mainnet` in case of `mainnet', () => {
-      expect(utils.toBCHAddressNetwork(Network.Mainnet)).toEqual(bchaddr.Network.Mainnet)
+      expect(Utils.toBCHAddressNetwork(Network.Mainnet)).toEqual(bchaddr.Network.Mainnet)
     })
     it('returns `bchaddr.Network.Testnet` in case of `testnet', () => {
-      expect(utils.toBCHAddressNetwork(Network.Testnet)).toEqual(bchaddr.Network.Testnet)
+      expect(Utils.toBCHAddressNetwork(Network.Testnet)).toEqual(bchaddr.Network.Testnet)
     })
   })
 
   describe('getBalance', () => {
     it('confirmed + unconfirmed', async () => {
       const address = 'qr20g55jd7x3dalp4qxjfgfvda0nwr8cfccrgxd0dw'
-      const balance = await utils.getBalance({ address, haskoinUrl })
+      const balance = await Utils.getBalance({ address, haskoinUrl })
       expect(balance.length).toEqual(1)
       expect(balance[0]?.asset).toEqual(AssetBCH)
       expect(balance[0]?.amount.amount().toString()).toEqual('300000000')
@@ -70,7 +70,7 @@ describe('Bitcoin Cash Utils Test', () => {
     it('returns txHash', async () => {
       const txHex = '0xdead'
 
-      const txHash = await utils.broadcastTx({ txHex, haskoinUrl })
+      const txHash = await Utils.broadcastTx({ txHex, haskoinUrl })
       expect(txHash).toEqual('mock-txid-haskoin')
     })
   })