+++ packages/xchain-doge/src/const.ts
@@ -1,3 +1,5 @@
+import { Asset, Chain } from '@xchainjs/xchain-util'
+
 /**
  * Minimum transaction fee
  * 100000 satoshi/kB (similar to current `minrelaytxfee`)
@@ -7,3 +9,17 @@ export const MIN_TX_FEE = 100000
 export const DOGE_DECIMAL = 8
 export const LOWER_FEE_BOUND = 40_000
 export const UPPER_FEE_BOUND = 20_000_000
+
+/**
+ * Chain identifier for Dogecoin
+ *
+ */
+export const DOGEChain: Chain = 'DOGE'
+
+/**
+ * Base "chain" asset on dogecoin
+ *
+ * Based on definition in Thorchain `common`
+ * @see https://gitlab.com/thorchain/thornode/-/blob/master/common/asset.go#L12-24
+ */
+export const AssetDOGE: Asset = { chain: DOGEChain, symbol: 'DOGE', ticker: 'DOGE', synth: false }