+++ packages/xchain-thorchain-query/__tests__/thorchain-query.test.ts
@@ -1,14 +1,10 @@
-import {
-  AssetAtom,
-  AssetBNB,
-  AssetBTC,
-  AssetETH,
-  AssetLTC,
-  AssetRuneNative,
-  assetAmount,
-  assetFromStringEx,
-  assetToBase,
-} from '@xchainjs/xchain-util'
+import { AssetBNB } from '@xchainjs/xchain-binance'
+import { AssetBTC } from '@xchainjs/xchain-bitcoin'
+import { AssetATOM } from '@xchainjs/xchain-cosmos'
+import { AssetETH } from '@xchainjs/xchain-ethereum'
+import { AssetLTC } from '@xchainjs/xchain-litecoin'
+import { AssetRuneNative } from '@xchainjs/xchain-thorchain'
+import { assetAmount, assetFromStringEx, assetToBase } from '@xchainjs/xchain-util'
 import { BigNumber } from 'bignumber.js'
 
 import mockMidgardApi from '../__mocks__/midgard-api'
@@ -254,10 +250,10 @@ describe('Thorchain-query tests', () => {
   it(`Should check assets match asset pools`, async () => {
     const assetPoolEthUsdc = await thorchainQuery.thorchainCache.getPoolForAsset(assetEthUSDC)
     const assetPoolAvaxUsdc = await thorchainQuery.thorchainCache.getPoolForAsset(assetAVAXUSDC)
-    const assetPoolGaia = await thorchainQuery.thorchainCache.getPoolForAsset(AssetAtom)
+    const assetPoolGaia = await thorchainQuery.thorchainCache.getPoolForAsset(AssetATOM)
     expect(assetPoolEthUsdc.asset).toEqual(assetEthUSDC)
     expect(assetPoolAvaxUsdc.asset).toEqual(assetAVAXUSDC)
-    expect(assetPoolGaia.asset).toEqual(AssetAtom)
+    expect(assetPoolGaia.asset).toEqual(AssetATOM)
   })
   it('Should construct the correct memo for BTC->BUSD swap', async () => {
     const swapParams: EstimateSwapParams = {