+++ packages/xchain-thorchain-query/src/thorchain-query.ts
@@ -1,10 +1,10 @@
+import { AssetBNB, BNBChain } from '@xchainjs/xchain-binance'
+import { AssetBTC } from '@xchainjs/xchain-bitcoin'
+import { GAIAChain } from '@xchainjs/xchain-cosmos'
+import { AssetRuneNative, THORChain, isAssetRuneNative } from '@xchainjs/xchain-thorchain'
 import { LastBlock } from '@xchainjs/xchain-thornode'
 import {
   Asset,
-  AssetAtom,
-  AssetBNB,
-  AssetBTC,
-  AssetRuneNative,
   Chain,
   assetAmount,
   assetFromStringEx,
@@ -13,7 +13,6 @@ import {
   baseAmount,
   eqAsset,
   getContractAddressFromAsset,
-  isAssetRuneNative,
 } from '@xchainjs/xchain-util'
 import { BigNumber } from 'bignumber.js'
 
@@ -288,7 +287,7 @@ export class ThorchainQuery {
     const contractAddress = getContractAddressFromAsset(asset)
     if (contractAddress && contractAddress.length > 5) {
       const abrev = contractAddress.substring(contractAddress.length - 5)
-      const sep = asset.chain !== Chain.THORChain && asset.synth ? '/' : '.'
+      const sep = asset.chain !== THORChain && asset.synth ? '/' : '.'
       return `${asset.chain}${sep}${asset.ticker}-${abrev}`
     }
     return assetToString(asset)
@@ -320,7 +319,7 @@ export class ThorchainQuery {
   // private async validateAffiliateAddress(affiliateAddress: string) {
   //   // Affiliate address should be THORName or THORAddress
   //   if (affiliateAddress.length > 0) {
-  //     const isValidThorchainAddress = this.clients[Chain.THORChain].validateAddress(affiliateAddress)
+  //     const isValidThorchainAddress = this.clients[THORChain].validateAddress(affiliateAddress)
   //     const isValidThorname = await this.isThorname(affiliateAddress)
   //     if (!(isValidThorchainAddress || isValidThorname))
   //       throw Error(`affiliateAddress ${affiliateAddress} is not a valid THOR address`)
@@ -445,11 +444,11 @@ export class ThorchainQuery {
     // RUNE, BNB and Synths have near instant finality, so no conf counting required. - need to make a BFT only case.
     if (
       isAssetRuneNative(inbound.asset) ||
-      inbound.asset.chain == AssetBNB.chain ||
-      inbound.asset.chain == AssetAtom.chain ||
+      inbound.asset.chain == BNBChain ||
+      inbound.asset.chain == GAIAChain ||
       inbound.asset.synth
     ) {
-      return this.chainAttributes[Chain.THORChain].avgBlockTimeInSecs
+      return this.chainAttributes[THORChain].avgBlockTimeInSecs
     }
     // Get the gas asset for the inbound.asset.chain
     const chainGasAsset = getChainAsset(inbound.asset.chain)
@@ -483,7 +482,7 @@ export class ThorchainQuery {
       .amount()
       .toNumber()
     const getScheduledOutboundValue = await this.thorchainCache.midgard.getScheduledOutboundValue()
-    const thorChainblocktime = this.chainAttributes[Chain.THORChain].avgBlockTimeInSecs // blocks required to confirm tx
+    const thorChainblocktime = this.chainAttributes[THORChain].avgBlockTimeInSecs // blocks required to confirm tx
     // If asset is equal to Rune set runeValue as outbound amount else set it to the asset's value in rune
     const runeValue = await this.thorchainCache.convert(outboundAmount, AssetRuneNative)
     // Check rune value amount