+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.0.20.7 (2022-12-27)
+
+## Add
+
+- Add `AssetBTC` and `BTCChain` definition
+
+## Update
+
+- Bump `xchain-client@13.5.0`
+
 # v.0.20.6 (2022-11-24)
 
 ## Update