+++ packages/xchain-bitcoincash/CHANGELOG.md
@@ -1,15 +1,24 @@
+# v.0.15.6 (2022-12-27)
+
+## Add
+
+- Add `AssetBCH` and `BCHChain` definition
+
+## Update
+
+- Bump `xchain-client@13.5.0`
+
 # v.0.15.5 (2022-12-13)
 
 ## Update
 
-- reverted `customRequestHeaders` to `BroadcastTxParams` 
-
+- reverted `customRequestHeaders` to `BroadcastTxParams`
 
 # v.0.15.4 (2022-11-24)
 
 ## Update
 
-- Added `customRequestHeaders` to `BroadcastTxParams` 
+- Added `customRequestHeaders` to `BroadcastTxParams`
 - Bumped `xchain-client`
 
 # v.0.15.3 (2022-10-14)