+++ packages/xchain-ethereum/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.0.27.6 (2022-12-27)
+
+## Add
+
+- Add `AssetETH` and `ETHChain` definition
+
+## Update
+
+- Bump `xchain-client@13.5.0`
+
 # v.0.27.5 (2022-12-17)
 
 ## Update