+++ packages/xchain-evm/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.0.1.3 (2022-12-27)
+
+## Add
+
+- Add `AssetAVAX` and `AVAXChain` definitions
+
+## Update
+
+- Bump `xchain-client@13.5.0`
+
 # v.0.1.2 (2022-12-17)
 
 ## Update